Refactor: export 방식 수정
