--- src/components/InfoWindow/index.ts
@@ -1,2 +1,0 @@
-import InfoWindow from './InfoWindow.vue';
-export { InfoWindow };
