--- src/components/KakaoMap/index.ts
@@ -1,2 +1,0 @@
-import KakaoMap from './KakaoMap.vue';
-export { KakaoMap };
