--- src/components/MapMarker/index.ts
@@ -1,3 +1,0 @@
-import MapMarker from './MapMarker.vue';
-export * from './types';
-export { MapMarker };
