+++ src/components/index.ts
@@ -1,3 +1,6 @@
-export * from './InfoWindow';
-export * from './KakaoMap';
-export * from './MapMarker';
+import KakaoMap from './KakaoMap/KakaoMap.vue';
+import MapMarker from './MapMarker/MapMarker.vue';
+import InfoWindow from './InfoWindow/InfoWindow.vue';
+export * from './MapMarker/types';
+
+export { KakaoMap, MapMarker, InfoWindow };
